requirements.txt
-i https://pypi.org/simple
alembic==1.16.5; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
bcrypt==4.3.0; python_version >= '3.8'
click==8.3.0; python_version >= '3.10'
fastapi==0.116.1; python_version >= '3.8'
greenlet==3.2.4; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.3; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
passlib==1.7.4
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.43; python_version >= '3.7'
starlette==0.47.3; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn==0.37.0; python_version >= '3.9'
uvicorn-worker==0.4.0; python_version >= '3.9'
